+++ src/main/java/ar/fiuba/tdd/tp/nikoligames/view/cells/numbercell/ClueCellView.java
@@ -3,7 +3,7 @@ package ar.fiuba.tdd.tp.nikoligames.view.cells.numbercell;
 import ar.fiuba.tdd.tp.nikoligames.view.ColorSet;
 import ar.fiuba.tdd.tp.nikoligames.view.cells.CellView;
 import ar.fiuba.tdd.tp.nikoligames.view.listeners.selectvaluelistener.SelectComponentCellListener;
-import ar.fiuba.tdd.tp.nikoligames.view.viewcontroller.AbstractSelectValueController;
+import ar.fiuba.tdd.tp.nikoligames.view.viewcontroller.SelectValueController;
 
 import java.awt.*;
 
@@ -17,14 +17,14 @@ public class ClueCellView extends CellView {
     private Color foregroundColor = ColorSet.CLUE;
 
 
-    public ClueCellView(int clue, AbstractSelectValueController selectValueController) {
+    public ClueCellView(String clue, SelectValueController selectValueController) {
         setBackground(backgroundColor);
         setForeground(foregroundColor);
-        this.setText(Integer.toString(clue));
+        this.setText(clue);
         registerMouseListener(selectValueController);
     }
 
-    private void registerMouseListener(AbstractSelectValueController selectValueController) {
+    private void registerMouseListener(SelectValueController selectValueController) {
         this.addMouseListener(new SelectComponentCellListener(selectValueController));
     }
