+++ src/main/java/ar/fiuba/tdd/tp/nikoligames/view/cells/numbercell/ValueClearerCell.java
@@ -3,7 +3,7 @@ package ar.fiuba.tdd.tp.nikoligames.view.cells.numbercell;
 import ar.fiuba.tdd.tp.nikoligames.view.ColorSet;
 import ar.fiuba.tdd.tp.nikoligames.view.cells.CellView;
 import ar.fiuba.tdd.tp.nikoligames.view.listeners.selectvaluelistener.ClearCellListener;
-import ar.fiuba.tdd.tp.nikoligames.view.viewcontroller.AbstractSelectValueController;
+import ar.fiuba.tdd.tp.nikoligames.view.viewcontroller.SelectValueController;
 
 import java.awt.*;
 
@@ -15,14 +15,14 @@ public class ValueClearerCell extends CellView {
     private Color backgroundColor = ColorSet.NORMAL_BACKGROUND;
     private Color foregroundColor = ColorSet.NUMBER;
 
-    public ValueClearerCell(AbstractSelectValueController selectValueController) {
+    public ValueClearerCell(SelectValueController selectValueController) {
         setBackground(backgroundColor);
         setForeground(foregroundColor);
         setText("X");
         registerMouseListener(selectValueController);
     }
 
-    protected void registerMouseListener(AbstractSelectValueController selectValueController) {
+    protected void registerMouseListener(SelectValueController selectValueController) {
         this.addMouseListener(new ClearCellListener(selectValueController));
     }
