+++ src/main/java/ar/fiuba/tdd/tp/nikoligames/view/gamebuttons/factory/BasicGroupButtonFactory.java
@@ -1,5 +1,6 @@
 package ar.fiuba.tdd.tp.nikoligames.view.gamebuttons.factory;
 
+import ar.fiuba.tdd.tp.nikoligames.engine.model.game.Game;
 import ar.fiuba.tdd.tp.nikoligames.engine.model.game.GameImplementation;
 import ar.fiuba.tdd.tp.nikoligames.view.gamebuttons.CheckGameButton;
 
@@ -15,7 +16,7 @@ import java.awt.*;
  */
 public class BasicGroupButtonFactory implements GroupButtonFactory {
 
-    public Component makeGroupButton(GameImplementation game) {
+    public Component makeGroupButton(Game game) {
         JPanel panel = new JPanel();
         JButton checkGame = new CheckGameButton(game);
         panel.add(checkGame);
