+++ src/main/java/ar/fiuba/tdd/tp/nikoligames/view/listeners/selectvaluelistener/SelectEditableCellListener.java
@@ -1,7 +1,7 @@
 package ar.fiuba.tdd.tp.nikoligames.view.listeners.selectvaluelistener;
 
 import ar.fiuba.tdd.tp.nikoligames.view.cells.numbercell.AbstractEditableNumberCell;
-import ar.fiuba.tdd.tp.nikoligames.view.viewcontroller.AbstractSelectValueController;
+import ar.fiuba.tdd.tp.nikoligames.view.viewcontroller.SelectValueController;
 
 import java.awt.event.MouseEvent;
 
@@ -14,7 +14,7 @@ import java.awt.event.MouseEvent;
  */
 public class SelectEditableCellListener extends AbstractSelectValueListener {
 
-    public SelectEditableCellListener(AbstractSelectValueController selectValueController) {
+    public SelectEditableCellListener(SelectValueController selectValueController) {
         super(selectValueController);
     }
