--- src/main/java/ar/fiuba/tdd/tp/nikoligames/view/parentview/factory/AbstractFactoryGameView.java
@@ -1,14 +1,0 @@
-package ar.fiuba.tdd.tp.nikoligames.view.parentview.factory;
-
-import ar.fiuba.tdd.tp.nikoligames.engine.model.game.GameImplementation;
-import ar.fiuba.tdd.tp.nikoligames.view.parentview.GameView;
-
-/**
- * Responsabilidades:
- * 1. Encapsular la creación del componente GameView
- * Patrón:
- * 1. Abstract Factory
- */
-public interface AbstractFactoryGameView {
-    GameView createDefaultGameView(GameImplementation game);
-}
